requirements.txt
joblib==1.2.0
scikit-learn==1.5.2
pandas==2.2.2
Django==5.1.2
djangorestframework==3.15.2
matplotlib==3.9.2
numpy==2.0.2
shap==0.46.0
lime==0.2.0.1